Changed some stuff to make it look pretty
Added ability to create a post, would be nice if we can get it to show which image we are uploading before we submit it. Also gotta make that page look nicer. Also fixed the profile, was trying to display our own posts on there but it was breaking so I just commented those functions out.

Cleaned up the way a post is displayed:

Fixed up some fonts and sizing, as well as commented out the paper plane icon since we don't make use of it in our app. Also added a divider so that you can distinguish between each post.