Adjusted comments...
I adujsted the way a comment is inputted.

Changed the floatingaction button to a "Post" button that is inside the text field just like instagram.

Also adjusted the overflow of the comment. It no longer goes off screen.

Added a popup window which will allow you to edit your comment and delete it if you are the user who made it.

STILL NEED TO IMPLEMENT:

Need to make the edit button create another pop up in which we can edit the comment, instead of using the main textfield from the comments page.

Also need to implement the user posts in user profiles.